backend/requirements.txt
# Backend Dependencies for Local Producer Web Application

# Core Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0

# MongoDB Database Driver
pymongo==4.6.1

# SMS Integration
twilio==8.10.3

# Authentication & Security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Environment Management
python-dotenv==1.0.0

# HTTP Client & Utilities
requests==2.31.0

# Input Validation
jsonschema==4.20.0

# Rate Limiting
Flask-Limiter==3.5.0

# Logging
structlog==23.2.0

# Testing Dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0

# Development Dependencies
python-dateutil==2.8.2